Entity changes
data type  of quantity changed from Integer to Double

+++ honeycakes-in/src/main/java/com/honeycakesin/dto/VendorItemsDto.java
@@ -18,7 +18,7 @@ public class VendorItemsDto {
 	
 	Double itemPrice;
 	
-	Integer quantitySlab;
+	Double quantitySlab;
 	
 	AvailabilityStatus availabilityStatus;
